+++ HISTORY.rst
@@ -18,6 +18,7 @@ Full Change List
 - Added command to rebuild variant summary materialized view ``rebuild_variant_summary``.
 - Adding link-out to PubMed with gene symbol and phenotype term names.
 - Improving existing link-out to Entrez Gene if the Entrez gene ID is known.
+- Adding ``ImportInfo`` to django admin.
 
 -------
 v0.14.7